Use PVertex via RecoContainer in the workflows
