+++ Detectors/GlobalTrackingWorkflow/readers/include/GlobalTrackingWorkflowReaders/PrimaryVertexReaderSpec.h
@@ -13,59 +13,13 @@
 #ifndef O2_PRIMARY_VERTEXREADER
 #define O2_PRIMARY_VERTEXREADER
 
-#include "TFile.h"
-#include "TTree.h"
-
 #include "Framework/DataProcessorSpec.h"
 #include "Framework/Task.h"
-
-#include "CommonDataFormat/TimeStamp.h"
-#include "ReconstructionDataFormats/VtxTrackIndex.h"
-#include "ReconstructionDataFormats/VtxTrackRef.h"
-#include "ReconstructionDataFormats/PrimaryVertex.h"
-#include "SimulationDataFormat/MCEventLabel.h"
 
 namespace o2
 {
 namespace vertexing
 {
-// read primary vertices produces by the o2-primary-vertexing-workflow
-
-class PrimaryVertexReader : public o2::framework::Task
-{
-  using Label = o2::MCEventLabel;
-  using V2TRef = o2::dataformats::VtxTrackRef;
-  using PVertex = o2::dataformats::PrimaryVertex;
-  using GIndex = o2::dataformats::VtxTrackIndex;
-
- public:
-  PrimaryVertexReader(bool useMC) : mUseMC(useMC) {}
-  ~PrimaryVertexReader() override = default;
-  void init(o2::framework::InitContext& ic) final;
-  void run(o2::framework::ProcessingContext& pc) final;
-
- protected:
-  void connectTree();
-
-  bool mVerbose = false;
-  bool mUseMC = false;
-
-  std::vector<PVertex> mVertices, *mVerticesPtr = &mVertices;
-  std::vector<Label> mLabels, *mLabelsPtr = &mLabels;
-  std::vector<V2TRef> mPV2MatchIdxRef, *mPV2MatchIdxRefPtr = &mPV2MatchIdxRef;
-  std::vector<GIndex> mPV2MatchIdx, *mPV2MatchIdxPtr = &mPV2MatchIdx;
-
-  std::unique_ptr<TFile> mFile;
-  std::unique_ptr<TTree> mTree;
-  std::string mFileName = "";
-  std::string mFileNameMatches = "";
-  std::string mVertexTreeName = "o2sim";
-  std::string mVertexBranchName = "PrimaryVertex";
-  std::string mVertexTrackIDsBranchName = "PVTrackIndices";
-  std::string mVertex2TrackIDRefsBranchName = "PV2TrackRefs";
-  std::string mVertexLabelsBranchName = "PVMCTruth";
-};
-
 /// create a processor spec
 /// read primary vertex data from a root file
 o2::framework::DataProcessorSpec getPrimaryVertexReaderSpec(bool useMC);
