+++ Detectors/Vertexing/src/SVertexer.cxx
@@ -30,15 +30,11 @@ using TrackITS = o2::its::TrackITS;
 using TrackTPC = o2::tpc::TrackTPC;
 
 //__________________________________________________________________
-void SVertexer::process(const gsl::span<const PVertex>& vertices,           // primary vertices
-                        const gsl::span<const GIndex>& trackIndex,          // Global ID's for associated tracks
-                        const gsl::span<const VRef>& vtxRefs,               // references from vertex to these track IDs
-                        const o2::globaltracking::RecoContainer& recoTracks // accessor to various tracks
-)
+void SVertexer::process(const o2::globaltracking::RecoContainer& recoData) // accessor to various reconstrucred data types
 {
   updateTimeDependentParams(); // TODO RS: strictly speaking, one should do this only in case of the CCDB objects update
-  mPVertices = vertices;
-  buildT2V(trackIndex, vtxRefs, recoTracks);
+  mPVertices = recoData.getPrimaryVertices<PVertex>();
+  buildT2V(recoData); // build track->vertex refs from vertex->track (if other workflow will need this, consider producing a message in the VertexTrackMatcher)
   int ntrP = mTracksPool[POS].size(), ntrN = mTracksPool[NEG].size(), iThread = 0;
   mV0sTmp[0].clear();
   mCascadesTmp[0].clear();
@@ -148,15 +144,15 @@ void SVertexer::setupThreads()
 }
 
 //__________________________________________________________________
-void SVertexer::buildT2V(const gsl::span<const GIndex>& trackIndex,           // Global ID's for associated tracks
-                         const gsl::span<const VRef>& vtxRefs,                // references from vertex to these track IDs
-                         const o2::globaltracking::RecoContainer& recoTracks) // accessor to various tracks
+void SVertexer::buildT2V(const o2::globaltracking::RecoContainer& recoData) // accessor to various tracks
 {
   // build track->vertices from vertices->tracks, rejecting vertex contributors
+  auto trackIndex = recoData.getPrimaryVertexMatchedTracks<GIndex>(); // Global ID's for associated tracks
+  auto vtxRefs = recoData.getPrimaryVertexMatchedTrackRefs<VRef>();   // references from vertex to these track IDs
 
   // track selector: at the moment reject prompt tracks contributing to vertex fit and unconstrained TPC tracks
   auto selTrack = [&](GIndex gid) {
-    return (gid.isPVContributor() || !recoTracks.isTrackSourceLoaded(gid.getSource())) ? false : true;
+    return (gid.isPVContributor() || !recoData.isTrackSourceLoaded(gid.getSource())) ? false : true;
   };
 
   std::unordered_map<GIndex, std::pair<int, int>> tmap;
@@ -183,7 +179,7 @@ void SVertexer::buildT2V(const gsl::span<const GIndex>& trackIndex,           //
           continue;
         }
       }
-      const auto& trc = recoTracks.getTrack(tvid);
+      const auto& trc = recoData.getTrack(tvid);
       int posneg = trc.getSign() < 0 ? 1 : 0;
       mTracksPool[posneg].emplace_back(TrackCand{trc, tvid, {iv, iv}});
       if (tvid.isAmbiguous()) { // track attached to >1 vertex, remember that it was already processed
